requirements.txt
ttkthemes==3.2.2
pillow==10.2.0 